Add files via upload
This repository contains code to find the middle node of a singly linked list using two approaches. The first approach calculates the list's length in the first pass and identifies the middle node in the second.  The second is the two-pointer technique, where a slow pointer and a fast pointer traverse the list, with the slow pointer identifying the middle node as the fast pointer reaches the end,  providing two efficient solutions for this fundamental problem.